\name{factorial}
\alias{factorial}
\title{factorial de un numero}

\description{ funcion que hace e factorial del numero dado}

\usage{
factorial(x)
}

\arguments{
  \item{x}{numero del que hara el factorial

}

}

\value{
Devuelve el factorial

\author{
Maria Tosar
}



\examples{
factorial(4)
fctorial(3)
## The function is currently defined as
function (x)
{
    prod(1:x)
  }
}
