\name{resta.dos.números}
\alias{resta.dos.números}
\title{Resta Dos Numeros}

\description{ funcion que resta el valor de dos numeros dados}

\usage{
resta.dos.números(x, y)
}

\arguments{
  \item{x}{
Primer numero del que se restara el segundo
}
  \item{y}{
Numero que se resta al primero
}
}

\value{
Devuelve la diferencia de los dos numeros
}

\author{
Maria Tosar
}



\examples{
resta.dos.numeros(3,1)
resta.dos.numeros(1,4)
## The function is currently defined as
function (x, y) 
{
    (x - y)
  }
}
