\name{suma.dos.números}
\alias{suma.dos.números}
\title{Duma Dos Numeros}

\description{ funcion que suma el valor de dos numeros dados, separados por comas}

\usage{
suma.dos.números(x, y)
}
\arguments{
  \item{x}{
Primer numero al que se le sumara el segundo
}
  \item{y}{
Numero que se suma al primero
}
}

\value{
Devuelve la suma de los dos numeros
}

\author{
Maria Tosar
}

