\name{suma.dos.números}
\alias{suma.dos.números}
\title{Duma Dos Numeros}

\description{ funcion que suma el valor de dos numeros dados}

\usage{
suma.dos.números(x, y)
}

\arguments{
  \item{x}{
Primer numero al que se le sumara el segundo
}
  \item{y}{
Numero que se suma al primero
}
}

\value{
Devuelve la suma de los dos numeros
}

\author{
Maria Tosar
}



\examples{
suma.dos.numeros(3,1)
suma.dos.numeros(1,4)
## The function is currently defined as
function (x, y) 
{
    (x + y)
  }
}
